refactor iOS
